Refactoring onboarding steps
Addressing comments
