+++ client/src/App.js
@@ -7,10 +7,8 @@ import SignUp from "./pages/SignUp";
 import Home from "./pages/Home";
 import CalendarPage from "./pages/CalendarPage";
 import { UserContextProvider } from "./globals/UserContext";
-import ProfileSetting from "./pages/ProfileSetting";
-import CalendarConfirm from "./pages/CalendarConfirm";
-import UserAvailability from "./pages/UserAvailability";
 import StripeCheckout from "./pages/StripeCheckout";
+import Onboarding from "./pages/Onboarding";
 import "./App.css";
 
 function App() {
@@ -22,9 +20,7 @@ function App() {
           <Route path="/signup" component={SignUp} />
           <Route path="/calendar" component={CalendarPage} />
           <Route exact path={["/", "/home", "/home/:page"]} component={Home} />
-          <Route exact path="/profile-settings" component={ProfileSetting} />
-          <Route exact path="/calendar-confirm" component={CalendarConfirm} />
-          <Route exact path="/availability" component={UserAvailability} />
+          <Route exact path={["/onboarding/profile-settings", "/onboarding/calendar-confirm", "/onboarding/availability"]} component={Onboarding} />
           <Route exact path="/checkout" component={StripeCheckout} />
         </BrowserRouter>
       </UserContextProvider>
