+++ client/src/pages/Onboarding.js
@@ -1,12 +1,14 @@
 import React, { useState } from "react";
+import { Route, Switch, Redirect } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import MobileStepper from '@material-ui/core/MobileStepper';
-import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
 import logo from '../images/logo.png';
 import { theme } from "../themes/theme";
-import { Link, useHistory } from 'react-router-dom';
+import ProfileSetting from "../components/onboarding/ProfileSetting";
+import CalendarConfirm from "../components/onboarding/CalendarConfirm";
+import UserAvailability from "../components/onboarding/UserAvailability";
 
 
 const useStyles = makeStyles({
@@ -58,7 +60,7 @@ function ProgressMobileStepper({activeStep}) {
     return (
       <MobileStepper
         variant="progress"
-        steps={5}
+        steps={4}
         position="static"
         activeStep={activeStep}
         className={classes.stepper}
@@ -68,44 +70,9 @@ function ProgressMobileStepper({activeStep}) {
     );
 }
 
-function Onboarding({children,title,activeStep,path,url}) {
+function OnboardingLayout({children,title,activeStep}) {
 
     const classes = useStyles();
-    
-    const history = useHistory();
-
-    const [urlStore, setUrlStore] = useState("");
-
-    const handleContinueClick = () => {
-        let status;
-        if (activeStep === 1){
-            if (url === "") return alert('empty url');
-            fetch(`/user/${url}/event-type`, {
-                method: "GET",
-                headers: {
-                    "Content-Type": "application/json"
-                }
-            })
-                .then(res => {
-                    status = res.status;
-                    if (status < 500) return res.json();
-                    else throw Error("Server error");
-                })
-                .then(res => {
-                    if (status === 200) {
-                        if (res.unique === true) {
-                            setUrlStore(url);
-                            return history.push(path);
-                        };
-                        return alert("Url exist")
-                    };
-                    throw Error("Failed to check");                
-                })
-                .catch(err => {
-                    alert(err.message);
-                });
-        }
-    }
 
     return (
 
@@ -124,24 +91,41 @@ function Onboarding({children,title,activeStep,path,url}) {
                     </Grid>
                 </Grid>
 
-                <Grid item container xs={12} className={classes.body} direction="column" spacing={4} justify="center">
-                    {children}
-                </Grid>
-
-                <Grid item container xs={12} justify="center"> 
-                    <Grid item xs={2}>
-                        <Button color="primary" variant="contained" fullWidth={true} className={classes.button}
-                        onClick={handleContinueClick}>
-                            <Link className={classes.link}>
-                                <Typography variant="body2">{activeStep === 3? "Finish":"Continue"}</Typography>
-                            </Link>
-                        </Button>
-                    </Grid>
-                </Grid>
+                {children}
             </Grid>
         </Grid>
     )
-
 };
+
+function renderSwitch(currPage, setters) {
+    switch(currPage) {
+        case "/onboarding/profile-settings":
+            return <ProfileSetting setters={setters}/>        
+        case "/onboarding/calendar-confirm":
+            return <CalendarConfirm setters={setters}/>
+        case "/onboarding/availability":
+            return <UserAvailability setters={setters}/>
+        default:
+            return <ProfileSetting  setters={setters}/>    
+    }
+}
+
+function Onboarding(props) {
+
+    const [activeStep, setActiveStep] = useState(1)
+    const [title, setTitle] = useState("Welcome to CalendApp!")
+    const currPage = props.location.pathname
+    
+    const setters = {
+        setActiveStep:setActiveStep,
+        setTitle:setTitle,
+    }
+
+    return (
+            <OnboardingLayout title={title} activeStep={activeStep}>
+                {renderSwitch(currPage, setters)}
+            </OnboardingLayout>
+        )
+}
 
 export default Onboarding
